+++ cluster/leader_client.go
@@ -107,7 +107,7 @@ func getURL(address string) string {
 	cfg := config.Get().(*conf.Configuration)
 
 	params := url.Values{}
-	params.Add(ClusterParam, "true")
+	params.Add(GETParam, "true")
 	// TODO выпилить instance_uuid ?
 	params.Add(utils.ModuleNameGetParamKey, cfg.ModuleName)
 	params.Add(utils.InstanceUuidGetParamKey, "9d89354b-c728-4b48-b002-a7d3b229f151")